requirements.txt
fastapi
uvicorn
python-docx
PyExifTool
PyPDF2
Pillow
python-multipart
python-dotenv